# docs/swagger.yaml
definitions:
...
  controller.ErrorResponse:
    properties:
      msg:
        type: string
    type: object
...
  dto.WishlistItemRequest:
    properties:
      product_id:
        example: 5c9f8f9e-7c1f-4b9c-8c1f-7c1f4b9c8c1f
        type: string
    type: object
  model.Category:
    properties:
      created_at:
        type: string
      description:
        type: string
      display_order:
        type: integer
      id:
        type: string
      is_active:
        type: boolean
      is_deleted:
        type: boolean
      name:
        type: string
      parent_id:
        type: string
      product_count:
        type: integer
      slug:
        type: string
      updated_at:
        type: string
    type: object
...
  model.Product:
    properties:
      attributes:
        additionalProperties: {}
        type: object
      categories:
        items:
          $ref: '#/definitions/model.Category'
        type: array
      cost_price:
        type: number
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      is_deleted:
        type: boolean
      name:
        type: string
      price:
        type: number
      sale_price:
        type: number
      sku:
        type: string
      status:
        type: string
      stock_quantity:
        type: integer
      updated_at:
        type: string
    type: object
...
  model.WishlistItem:
    properties:
      added_at:
        type: string
      product:
        $ref: '#/definitions/model.Product'
      product_id:
        type: string
      user_id:
        type: string
    type: object
info:
  contact: {}
  description: Fiber go web framework based REST API boilerplate
  title: Fiber Go API
  version: "1.0"
paths:
...
  /api/v1/wishlist:
    get:
      consumes:
      - application/json
      description: Get all items in the user's wishlist
      produces:
      - application/json
      responses:
        "200":
          description: Wishlist items
          schema:
            items:
              $ref: '#/definitions/model.WishlistItem'
            type: array
        "401":
          description: Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "500":
          description: Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: get user wishlist
      tags:
      - Wishlist
    post:
      consumes:
      - application/json
      description: Add a product to the user's wishlist
      parameters:
      - description: Product ID to add
        in: body
        name: wishlistItem
        required: true
        schema:
          $ref: '#/definitions/dto.WishlistItemRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema: {}
        "400":
          description: Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "401":
          description: Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "404":
          description: Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "500":
          description: Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: add product to wishlist
      tags:
      - Wishlist
  /api/v1/wishlist/{product_id}:
    delete:
      consumes:
      - application/json
      description: Remove a product from the user's wishlist
      parameters:
      - description: Product ID
        in: path
        name: product_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema: {}
        "400":
          description: Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "401":
          description: Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "404":
          description: Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "500":
          description: Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: remove product from wishlist
      tags:
      - Wishlist
  /api/v1/wishlist/check/{product_id}:
    get:
      consumes:
      - application/json
      description: Check if a product is in the user's wishlist
      parameters:
      - description: Product ID
        in: path
        name: product_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Result with in_wishlist field
          schema:
            additionalProperties:
              type: boolean
            type: object
        "400":
          description: Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "401":
          description: Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "500":
          description: Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: check wishlist item
      tags:
      - Wishlist
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"